requirements.txt
absl-py==1.3.0
addict==2.4.0
aiohttp==3.8.3
aiosignal==1.2.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
audioread==3.0.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==4.1.0
Bottleneck==1.3.4
brotlipy==0.7.0
cachetools==5.2.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.4
colorama==0.4.6
cryptography==37.0.1
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.1
ffmpeg-normalize==1.25.3
ffmpeg-progress-yield==0.3.0
filelock==3.8.0
fire==0.4.0
fonttools==4.25.0
frozenlist==1.3.1
fsspec==2022.10.0
future==0.18.2
gdown==4.5.3
google-auth==2.14.0
google-auth-oauthlib==0.4.6
grpcio==1.50.0
h5py==3.7.0
huggingface-hub==0.10.1
idna==3.3
importlib-metadata==5.0.0
ipykernel==6.9.1
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.4.0
julius==0.2.7
jupyter==1.0.0
jupyter-client==7.2.2
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.4.2
librosa==0.9.2
lightning-lite==1.8.0.post1
lightning-utilities==0.3.0
llvmlite==0.39.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.2
mistune==0.8.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
multidict==6.0.2
munkres==1.1.4
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.3.0
nest-asyncio==1.5.5
notebook==6.4.11
numba==0.56.4
numexpr==2.8.1
numpy==1.22.3
oauthlib==3.2.2
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pip==21.2.4
pooch==1.6.0
primePy==1.3
prometheus-client==0.13.1
prompt-toolkit==3.0.20
protobuf==3.19.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.11.2
pyloudnorm==0.1.0
pyOpenSSL==22.0.0
pyparsing==3.0.4
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-lightning==1.8.0.post1
pytz==2022.1
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.3.0
QtPy==2.0.1
regex==2022.10.31
requests==2.27.1
requests-oauthlib==1.3.1
resampy==0.4.2
rsa==4.9
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.1
Send2Trash==1.8.0
setuptools==61.2.0
sip==4.19.13
six==1.16.0
soundfile==0.11.0
soupsieve==2.3.1
stack-data==0.2.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.1.0
terminado==0.13.1
testpath==0.5.0
threadpoolctl==2.2.0
timm==0.6.11
tokenizers==0.13.1
torch==1.13.0+cu116
torch-audiomentations==0.11.0
torch-pitch-shift==1.2.2
torchaudio==0.13.0+cu116
torchaudio-augmentations==0.2.4
torchinfo==1.7.1
torchmetrics==0.10.2
torchsummary==1.5.1
torchvision==0.14.0+cu116
tornado==6.1
tqdm==4.64.1
traitlets==5.1.1
transformers==4.24.0
typing_extensions==4.1.1
urllib3==1.26.9
wavaugment==0.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
wheel==0.37.1
widgetsnbextension==3.5.2
yapf==0.32.0
yarl==1.8.1
zipp==3.10.0
